requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-extensions==4.1
djangorestframework==3.16.1
frozenlist==1.8.0
geoip2==5.2.0
gunicorn==23.0.0
idna==3.11
maxminddb==3.0.0
multidict==6.7.0
packaging==25.0
propcache==0.4.1
requests==2.32.5
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
whitenoise==6.11.0
yarl==1.22.0
